refactor: ProjectService가 ProjectRepository를 주입 받도록 변경
