+++ src/main/java/org/ject/support/domain/project/repository/ProjectRepository.java
@@ -3,5 +3,5 @@ package org.ject.support.domain.project.repository;
 import org.ject.support.domain.project.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project, Long> {
+public interface ProjectRepository extends JpaRepository<Project, Long>, ProjectQueryRepository {
 }
